backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
websockets==14.*
pydantic==2.*
pydantic-settings==2.*
httpx==0.28.*
numpy==2.*
faster-whisper==1.*
pyannote.audio==3.3.*
huggingface-hub<1.0
matplotlib
python-dotenv==1.*
